CRUD OK

+++ model/user_model.go
@@ -1,7 +1,7 @@
 package model
 
 type Users struct {
-	ID      int64  `json:"id" xorm:"id"`
+	ID      int    `json:"id" xorm:"id"`
 	Name    string `json:"name"`
 	Address string `json:"address"`
 	Email   string `json:"email"`
